Распределил тесты по сущностям в папке Tests
